+++ src/components/NewWorkout/Timer.tsx
