+++ src/pages/new-workout.tsx
@@ -1,7 +1,9 @@
-import Workout from '@components/Workout/NewWorkout';
+import Workout from '@components/NewWorkout/NewWorkout';
 import { WorkoutContextProvider } from '@contexts/WorkoutContext';
 
 const NewWorkout = () => {
+  // TODO: Reroute if not logged in
+
   return (
     <WorkoutContextProvider>
       <Workout />
